refactor: simplify build system and clean up condense files
