+++ tools/contact/src/datatype.ts
@@ -1,23 +1,5 @@
 import { type DataTypeImplementation } from "@patchwork/plugins";
-import { type AutomergeUrl } from "@automerge/automerge-repo";
-
-// SCHEMA
-
-export interface AnonymousContactDoc {
-  type: "anonymous";
-  color?: string; // HSL color string for user presence indicators
-}
-
-export interface RegisteredContactDoc {
-  type: "registered";
-  name: string;
-  avatarUrl?: AutomergeUrl;
-  color?: string; // HSL color string for user presence indicators
-}
-
-export type ContactDoc = AnonymousContactDoc | RegisteredContactDoc;
-
-// FUNCTIONS
+import type { ContactDoc } from "./types";
 
 export const init = (doc: ContactDoc) => {
   Object.assign(doc, {
