+++ tools/contact/vite.config.ts
@@ -1,23 +1,16 @@
-import react from "@vitejs/plugin-react";
 import { defineConfig } from "vite";
-import topLevelAwait from "vite-plugin-top-level-await";
-import cssInjectedByJsPlugin from "vite-plugin-css-injected-by-js";
 import tailwindcss from "@tailwindcss/vite";
+import react from "@vitejs/plugin-react";
+import cssInjectedByJsPlugin from "vite-plugin-css-injected-by-js";
+import externals from "@patchwork/bootloader/externals";
 
 export default defineConfig({
   base: "./",
-  plugins: [topLevelAwait(), react(), tailwindcss(), cssInjectedByJsPlugin()],
+  plugins: [react(), tailwindcss(), cssInjectedByJsPlugin()],
+
   build: {
     rollupOptions: {
-      external(id) {
-        // Don't externalize libraries that depend on react
-        // these need to share the same instance of react as the tool
-        if (id === "@patchwork/react" || id === "@patchwork/context-react")
-          return false;
-
-        // ... otherwise externalize all automerge-repo and @patchwork packages
-        return !!id.match(/^((@automerge\/automerge(-repo)?)|@patchwork\/.*)$/);
-      },
+      external: externals,
       input: "./src/tool.tsx",
       output: {
         format: "es",
